requirements.txt
Django==5.0.0
djangorestframework==3.15.0
psycopg2-binary==2.9.9
razorpay==1.4.2
gunicorn==21.2.0
django-widget-tweaks==1.5.0



